+++ packages/mdx/src/remark/transform.section.ts
